% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_character.R
\name{as_character}
\alias{as_character}
\title{Enhance the performance of as.character()}
\usage{
as_character(.x)
}
\arguments{
\item{.x}{A vector.}
}
\value{
A character vector with the same length as \code{.x}.
}
\description{
Enhance the performance of as.character()
}
\examples{
lifeexp \%>\%
  select(region)
lifeexp \%>\%
  mutate(
    region_chr = as_character(region),
    .keep = "used"
  )
}
