% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_type.R
\name{as_type}
\alias{as_type}
\alias{as_character}
\alias{as_numeric}
\title{Better type conversion}
\usage{
as_character(.x)

as_numeric(.x)
}
\arguments{
\item{.x}{A vector or matrix.}
}
\description{
It improves the performance of converting
several types of variables.
}
\examples{
lifeexp \%>\%
  dplyr::mutate(
    region_chr = as_character(region),
    region_num = as_numeric(region),
    region_fct = haven::as_factor(region),
    .keep = "used"
  )
}
