% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse}
\alias{browse}
\title{Browse columns like Stata's \code{browse} command}
\usage{
browse(.data, ...)
}
\arguments{
\item{.data}{A data frame}

\item{...}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> or
<\code{\link[=args_data_masking]{data-masking}}> Variables to browse.
\emph{This argument can be omitted}.}
}
\value{
A character vector of column names.
}
\description{
Browse columns like Stata's \code{browse} command
}
\examples{
browse(starwars)
browse(starwars, name:mass)
browse(starwars, s_match("*color"))
}
