% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.R
\name{cut}
\alias{cut}
\alias{cut_quantile}
\alias{cut_length}
\alias{cut_breaks}
\title{Cut a numeric vector into intervals.}
\usage{
cut_quantile(x, n = 4, labels = NULL)

cut_length(x, n = 4, labels = NULL)

cut_breaks(x, breaks = NULL, labels = NULL, right = TRUE)
}
\arguments{
\item{x}{A numeric vector.}

\item{n}{A positive integer. The number of intervals.}

\item{labels}{labels for the levels of the resulting category.  By default,
    labels are constructed using \code{"(a,b]"} interval notation.  If
    \code{labels = FALSE}, simple integer codes are returned instead of
    a factor.}

\item{breaks}{either a numeric vector of two or more unique cut points or a
    single number (greater than or equal to 2) giving the number of
    intervals into which \code{x} is to be cut.}

\item{right}{logical, indicating if the intervals should be closed on
    the right (and open on the left) or vice versa.}
}
\value{
Print the selected rows of the data.
Return the input data frame invisibly.
}
\description{
Cut a numeric vector into intervals.
}
\examples{
cut_quantile(1:20)
cut_length(1:20)
cut_breaks(1:20, breaks = c(5, 15))
}
