% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.R
\name{ds}
\alias{ds}
\title{List column names by tidyselect}
\usage{
ds(.data, ...)
}
\arguments{
\item{.data}{A data frame}

\item{...}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Variables to select.}
}
\value{
A character vector of column names.
}
\description{
List column names by tidyselect
}
\examples{
ds(starwars)
ds(starwars, homeworld:starships)
ds(starwars, ends_with("color"))
ds(starwars, where(is.character))
}
