% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_type.R
\name{s_type}
\alias{s_type}
\title{Get the abbreviated type of a vector.}
\usage{
s_type(.x)
}
\arguments{
\item{.x}{A vector.}
}
\value{
A character.
}
\description{
Get the abbreviated type of a vector.
}
\examples{
s_type(1:10)
s_type(letters)
}
