% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_tail_print.R
\name{head_tail_print}
\alias{head_tail_print}
\title{Print a long tibble with its first and last rows}
\usage{
head_tail_print(
  .data,
  n = 5,
  width = NULL,
  .head_row = FALSE,
  .tail_row = FALSE
)
}
\description{
Print a long tibble with its first and last rows
}
