% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-lifeexp.R
\docType{data}
\name{lifeexp}
\alias{lifeexp}
\title{Life expectancy data in 1998}
\format{
A tibble with 68 rows and 6 variables:
\describe{
\item{region}{Region}
\item{country}{Country}
\item{popgrowth}{Average annual population growth rate (\%)}
\item{lexp}{Life expectancy at birth (years)}
\item{gnppc}{GNP per capita}
\item{safewater}{Safe water (\%)}
}
}
\usage{
lifeexp
}
\description{
Imported from the example dataset \code{lifeexp.dta} in Stata.
This dataset contains variable labels and value labels.
}
\examples{
lifeexp
}
\keyword{datasets}
