% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_conversion.R
\name{na_conversion}
\alias{na_conversion}
\alias{na_if_value}
\alias{value_if_na}
\title{NA Conversion}
\usage{
na_if_value(.x, .value)

value_if_na(.x, .value)
}
\arguments{
\item{.x}{A vector.}

\item{.value}{A value to convert to NA.}
}
\description{
\itemize{
\item \code{na_if_value()} converts a value to NA. The same as \code{dplyr::na_if()}.
\item \code{value_if_na()} converts NA to a value. The same as \code{tidyr::replace_na()}.
}
}
\details{
Convert NAs easily.
}
\examples{
x <- c(1, 2, NA, 4)
na_if_value(x, 2)
value_if_na(x, 3)
}
