% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{names}
\alias{names}
\alias{ds}
\alias{ds_as_column}
\alias{names_as_column}
\title{List column names by tidyselect}
\usage{
ds(.data, ...)

ds_as_column(.data, ...)

names_as_column(.data, ...)
}
\arguments{
\item{.data}{A data frame}

\item{...}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> Variables to select.}
}
\value{
A character vector of column names.
}
\description{
List column names by tidyselect
}
\examples{
ds(starwars) # equivalent to names(starwars)

ds(starwars, homeworld:starships)

ds_as_column(starwars, ends_with("color"))

ds_as_column(starwars, where(is.character))


names_as_column(starwars)

names_as_column(starwars, ends_with("color"))

names_as_column(starwars, where(is.character))
}
\seealso{
\code{\link[=variables]{variables()}}, \code{\link[=codebook]{codebook()}}, \code{\link[=browse]{browse()}}
}
