% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{regress}
\alias{regress}
\alias{regress_coef}
\alias{regress_stat}
\alias{regress_data}
\alias{tidy_coef}
\title{Directly obtain the regression results}
\usage{
regress(.data, formula, model = "lm", digit = integer(0), format = "\%.3f")

regress_coef(
  .data,
  formula,
  model = "lm",
  digit = integer(0),
  format = "\%.3f"
)

regress_stat(.data, formula, model = "lm")

regress_data(.data, formula, model = "lm")

tidy_coef(model, digit = integer(0), format = "\%.3f")
}
\arguments{
\item{.data}{A data frame, data frame extension
(e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{formula}{A formula specifying the model.}

\item{model}{A character string specifying the model to be used.}

\item{digit}{A positive integer specifying the number of digits to be displayed.}

\item{format}{A character string specifying the format of the coefficients.
The default is \verb{\%.3f}, which means three decimal places.}
}
\value{
Return the regression results as a list or a tibble.
}
\description{
These functions are wrappers for the \code{broom} package to directly obtain the regression results.
The naming convention is \verb{regress_*} for raw formula and \verb{tidy_*} for the model object.
\itemize{
\item \code{regress()} returns the regression results as a list.
\item \code{regress_coef()} returns the coefficients.
\item \code{regress_stat()} returns the model statistics.
\item \code{regress_data()} returns the augmented data.
\item \code{tidy_coef()} returns the coefficients from the model object.
}
}
\examples{
regress(mtcars, mpg ~ wt + hp)
regress_coef(mtcars, mpg ~ wt + hp)

mtcars \%>\%
  dplyr::mutate(mpg20 = mpg > 20) \%>\%
  regress(mpg20 ~ wt + hp, model = "logit")
}
