% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_ggsave.R
\name{s_ggsave}
\alias{s_ggsave}
\title{Quickly save a ggplot to a file}
\usage{
s_ggsave(
  plot = ggplot2::last_plot(),
  filename,
  path = "wd",
  width = 7,
  height = 5,
  units = "in",
  dpi = 300,
  bg = "white"
)
}
\arguments{
\item{plot}{A ggplot object}

\item{filename}{A character string of the file name}

\item{path}{A character string of the file path. By default, it is the working directory.}

\item{width, height}{Plot size in units expressed by the \code{units} argument.
If not supplied, uses the size of the current graphics device.}

\item{units}{One of the following units in which the \code{width} and \code{height}
arguments are expressed: \code{"in"}, \code{"cm"}, \code{"mm"} or \code{"px"}.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{bg}{Background colour. If \code{NULL}, uses the \code{plot.background} fill value
from the plot theme.}
}
\description{
Quickly save a ggplot to a file
}
\examples{
gg <- s_plot(mtcars, mpg, wt)
# s_ggsave(gg, "plot.png")

# Alternatively, you can use the pipe operator
# mtcars \%>\%
#   s_plot(mpg, wt) \%>\%
#   s_ggsave("plot.png")
}
