% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_matches.R
\name{s_matches}
\alias{s_matches}
\title{Tidyselect columns in stata style}
\usage{
s_matches(string, ignore.case = TRUE, vars = NULL)
}
\arguments{
\item{string}{A character vector of column names
like selecting a varlist in Stata \code{keep} command.}

\item{ignore.case}{If \code{TRUE}, the default, ignores case when matching
names.}

\item{vars}{A character vector of variable names. If not supplied,
the variables are taken from the current selection context (as
established by functions like \code{select()} or \code{pivot_longer()}).}
}
\description{
\code{s_matches()} is a simpler version of \code{matches()}
in <\code{\link[=dplyr_tidy_select]{tidy-select}}>.
It is designed to select columns in a Stata-like style.

As \code{matches()}, \code{s_matches()} must be used
within a \emph{selecting} function, such as
\code{ds()} or
\code{dplyr::across()}.
}
\details{
\itemize{
\item \code{*} matches zero or more characters.
\item \code{~} matches one or more characters.
\item \verb{?} matches one character.
\item \verb{[a-h]} matches any character in the range from \code{a} to \code{h}.
\item \verb{[1-12]} matches any number in the range from \code{1} to \code{12}.
\item \code{var1-var5} is NOT used to select a range of variables.
Use \code{var1:var5} out of \code{s_matches()} instead.
}
}
\examples{
tb <- paste0("x", 1:15) \%>\%
  tibble::as_tibble() \%>\%
  tidyr::pivot_wider(names_from = value)
tb

# List variable names by ds()
ds(tb)

tb \%>\%
  ds(s_matches("x1*"))
tb \%>\%
  ds(s_matches("x1~"))
tb \%>\%
  ds(s_matches("*5"))
tb \%>\%
  ds(s_matches("?5"))
tb \%>\%
  ds(s_matches("x[9-15]"))

\dontrun{
tb \%>\%
  ds(s_matches("x1-x5"))
}

tb \%>\%
  ds(x1:x5)

}
