% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_print.R
\name{s_print}
\alias{s_print}
\title{Print a long tibble with its first and last rows}
\usage{
s_print(.data, n = 5, width = NULL, .head_row = FALSE, .tail_row = FALSE)
}
\description{
Print a long tibble with its first and last rows
}
