% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_select.R
\name{s_select}
\alias{s_select}
\title{Enhanced select() with data-masking}
\usage{
s_select(.data, ...)
}
\arguments{
\item{.data}{A data frame, data frame extension
(e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{...}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> or
<\code{\link[=args_data_masking]{data-masking}}> Variables to tabulate.

Both tidyselect (e.g., \code{starts_with()}) and
data masking (e.g., \code{x_sq = x^2}) are supported.
See examples below.}
}
\value{
A tibble
}
\description{
Enhanced select() with data-masking
}
\examples{
starwars \%>\%
 s_select()
starwars \%>\%
 s_select(name, height)
\dontrun{
starwars \%>\%
 dplyr::select(height > 50)
}
starwars \%>\%
 s_select(height > 50)
}
