% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_time.R
\name{s_time}
\alias{s_time}
\title{Time spent on a function}
\usage{
s_time(.x)
}
\arguments{
\item{.x}{The function}
}
\value{
The function output
}
\description{
Time spent on a function
}
\examples{
x <- seq_len(1e6)
mean(x) \%>\% s_time()
}
