% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_try.R
\name{s_try}
\alias{s_try}
\title{Try a function and return a logical}
\usage{
s_try(.fn)
}
\arguments{
\item{.fn}{A function call.}
}
\value{
A logical value.
}
\description{
Try a function and return a logical
}
