% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_type.R
\name{s_type}
\alias{s_type}
\alias{s_unit}
\title{The clean type of an object.}
\usage{
s_type(.x, .abbr = FALSE)

s_unit(.x)
}
\arguments{
\item{.x}{An object.}

\item{.abbr}{A logical. If \code{TRUE}, the abbreviation
of the type is returned.}
}
\value{
A character.
}
\description{
The clean type of an object.
}
\examples{
s_type(1:10)
s_type(letters)
}
