% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_seed.R
\name{set_seed}
\alias{set_seed}
\title{Set seeds with pipeline.}
\usage{
set_seed(.fn, seed)
}
\arguments{
\item{.fn}{A function call.}

\item{seed}{An integer for reproducibility.}
}
\value{
Return the function result with a seed set.
}
\description{
Attach a seed to a specific function.
}
\examples{
sample(1:10, 8) \%>\% 
  set_seed(123)

# When we re-run the code, we will get the same result.
sample(1:10, 8) \%>\%
 set_seed(123)
}
