% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepecial-print.R
\name{special-print}
\alias{special-print}
\alias{print_headtail}
\alias{print_interval}
\alias{slice_headtail}
\alias{slice_interval}
\title{Print a long tibble with selected rows}
\usage{
print_headtail(
  .data,
  n = 10,
  width = NULL,
  ...,
  max_extra_cols = NULL,
  max_footer_lines = NULL
)

print_interval(
  .data,
  n = 10,
  width = NULL,
  ...,
  max_extra_cols = NULL,
  max_footer_lines = NULL
)

slice_headtail(.data, n = 10)

slice_interval(.data, n = 10)
}
\arguments{
\item{.data}{A data frame, data frame extension
(e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{n}{A positive integer. The number of rows to show in
the head and in the tail.}

\item{width}{A positive integer. The width of the printed tibble.}

\item{...}{Passed on to \code{\link[=tbl_format_setup]{tbl_format_setup()}}.}

\item{max_extra_cols}{Number of extra columns to print abbreviated information for,
if the width is too small for the entire tibble. If \code{NULL},
the \code{max_extra_cols} \link[pillar:pillar_options]{option} is used.
The previously defined \code{n_extra} argument is soft-deprecated.}

\item{max_footer_lines}{Maximum number of footer lines. If \code{NULL},
the \code{max_footer_lines} \link[pillar:pillar_options]{option} is used.}
}
\value{
Print the selected rows of the data.
Return the input data frame invisibly.
}
\description{
\itemize{
\item \code{print_headtail()} - Print the head and tail rows of the data.
\item \code{print_interval()} - Print the selected rows with (almost) equal intervals.
\item \code{slice_headtail()} - Return the head and tail rows of the data.
\item \code{slice_interval()} - Return the selected rows with (almost) equal intervals.
}
}
\examples{
print_headtail(starwars)

print_interval(starwars)

slice_headtail(starwars)

slice_interval(starwars)
}
