% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat}
\alias{stat}
\alias{skew}
\alias{kurtosis}
\title{Skewness and kurtosis}
\usage{
skew(x, n)

kurtosis(x, n)
}
\arguments{
\item{x}{A numeric vector}

\item{n}{The number of observations}
}
\value{
A numeric value
}
\description{
Calculate skewness and kurtosis
}
