% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statart-package.R, R/statart.R
\docType{package}
\name{statart-package}
\alias{statart}
\alias{statart-package}
\title{statart: Tidy Functions with Stata Functionality}
\description{
To learn more about statart, start with the vignettes:
\code{browseVignettes(package = "statart")}

To learn more about statart, start with the vignettes:
\code{browseVignettes(package = "statart")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/socimh/statart}
  \item \url{https://socimh.github.io/statart/}
  \item Report bugs at \url{https://github.com/socimh/statart/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/socimh/statart}
  \item \url{https://socimh.github.io/statart/}
  \item Report bugs at \url{https://github.com/socimh/statart/issues}
}

}
\author{
\strong{Maintainer}: Minghong Shen \email{mhshenaa@connect.ust.hk}

}
\keyword{internal}
