% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab.R
\name{tab}
\alias{tab}
\alias{tab1}
\alias{tab2}
\alias{fre}
\title{Tabulate columns}
\usage{
tab(.data, ..., .desc = FALSE, .print = TRUE, .print_n = 10)

tab1(
  .data,
  ...,
  .desc = FALSE,
  .separate = FALSE,
  .print = TRUE,
  .print_n = 10
)

tab2(
  .data,
  row_var,
  col_var,
  .desc = FALSE,
  .print = TRUE,
  .print_n = 10,
  .print_width = NULL
)

fre(.data, ..., .desc = FALSE, .print = TRUE, .print_n = 10)
}
\arguments{
\item{.data}{The input .data (.data frame or tibble).}

\item{.desc}{A logical value.
If TRUE, sort the result in descending order by frequency.
Default is FALSE.}

\item{row_var}{The variable to tabulate.}

\item{col_var}{(Optional) The additional variable to cross-tabulate.}

\item{.quiet}{A logical value.
If TRUE, do not print the result. Default is FALSE.}
}
\value{
An object of the same type as \code{.data}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\itemize{
\item \code{fre()} tabulates the frequency and missing values of one or more variables.
\item \code{tab()} tabulates a single variable.
\item \code{tab1()} tabulates multiple variables at once.
\item \code{tab2()} cross-tabulates two variables.
}

Overall, \code{fre()} is a more general function
that is an enhanced version of \code{count()}.

However, \code{tab()} family functions could yield
different results that are sometimes more useful.
}
\examples{
# tab() the 8th column
tab(starwars, 8)
tab(starwars, sex)
tab(starwars, sex, .desc = TRUE)
tab(starwars, sex, gender)
tab(starwars, gender, sex)
}
