% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_coef.R
\name{tidy_coef}
\alias{tidy_coef}
\title{Directly obtain the regression results}
\usage{
tidy_coef(model, digit = integer(0), format = "\%.3f")
}
\arguments{
\item{model}{A character string specifying the model to be used.}

\item{digit}{A positive integer specifying the number of digits to be displayed.}

\item{format}{A character string specifying the format of the coefficients.
The default is \verb{\%.3f}, which means three decimal places.}
}
\value{
Return the regression results as a list or a tibble.
}
\description{
Given a model fit object (e.g. from \code{lm()} or \code{glm()}), \code{tidy_coef()} returns a tibble of coefficients.
}
\examples{
lm(mpg ~ wt + hp, data = mtcars) \%>\%
  tidy_coef()
}
\seealso{
\code{\link[=regress]{regress()}}
}
