% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{variables}
\alias{variables}
\alias{variables_search}
\title{Show the variable list of a dataset}
\usage{
variables(.data, ...)

variables_search(.data, string, ...)
}
\arguments{
\item{.data}{The input data (data frame or tibble).}

\item{...}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> or
<\code{\link[=args_data_masking]{data-masking}}> Variables to include in the codebook.
\emph{This argument can be omitted}.}

\item{string}{A string to search for in the variable names or labels.}
}
\value{
A tibble.
}
\description{
\itemize{
\item \code{variables()} works like the "Variables" tab in Stata.
\item \code{variables_search()} searches for variables by name or label.
}
}
\examples{
variables(lifeexp)

variables_search(lifeexp, "n")
}
\seealso{
\code{\link[=ds]{ds()}}, \code{\link[=codebook]{codebook()}}, \code{\link[=browse]{browse()}}
}
