% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{codebook}
\alias{codebook}
\title{View the codebook of a dataset or selected variables.}
\usage{
codebook(.data, ...)
}
\arguments{
\item{.data}{The input data (data frame or tibble).}

\item{...}{<\code{\link[=dplyr_tidy_select]{tidy-select}}> One or more
unquoted expressions separated by commas. Variable names can be
used as if they were positions in the data frame, so expressions
like x:y can be used to select a range of variables.
\emph{This argument can be omitted}.}
}
\value{
An object of the same type as \code{.data}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\code{codebook()} is a function to view the
codebook of a dataset or selected variables.
}
\examples{
codebook(starwars)
codebook(starwars, 1:4)
codebook(starwars, ends_with("color"))
codebook(starwars, where(is.numeric))
codebook(lifeexp)
}
