% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_type_abbr.R
\name{get_type_abbr}
\alias{get_type_abbr}
\title{Get the abbreviated type of a vector.}
\usage{
get_type_abbr(.x)
}
\arguments{
\item{.x}{A vector.}
}
\value{
A character.
}
\description{
Get the abbreviated type of a vector.
}
\examples{
get_type_abbr(1:10)
get_type_abbr(letters)
}
