% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-export.R
\name{import-export}
\alias{import-export}
\alias{read_data}
\alias{write_data}
\title{Import and export with various data formats}
\usage{
read_data(file, ...)

write_data(data, file, ...)
}
\arguments{
\item{file}{A file path.}

\item{...}{arguments passed to the appropriate function.}

\item{data}{A tibble.}
}
\value{
A tibble.
}
\description{
Import and export with various data formats
}
\examples{
ds(starwars)
}
