% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_try.R
\name{s_try}
\alias{s_try}
\title{Try a function and return a logical}
\usage{
s_try(.fn)
}
\arguments{
\item{.fn}{A function call.}
}
\value{
A logical value.
}
\description{
This function tries to evaluate a function,
and if the function throws an error, it returns FALSE.
Otherwise, it returns TRUE.
}
\examples{
s_try(mean(1:10))

# This gives a warning, but it is still working, so s_try() returns TRUE
s_try(mean(letters[1:10]))

# This throws an error, so s_try() returns FALSE
s_try(means(1:10))
}
