% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_tools.R
\name{s_type}
\alias{s_type}
\title{The clean type of an object.}
\usage{
s_type(.x, .abbr = FALSE)
}
\arguments{
\item{.x}{An object.}

\item{.full}{A logical. If \code{TRUE}, the full name of the type is returned.}
}
\value{
A character.
}
\description{
The clean type of an object.
}
\examples{
s_type(1:10)
s_type(letters) 
}
