% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ.R
\name{summ}
\alias{summ}
\title{Summarise columns}
\usage{
summ(.data, ..., .by = NULL, .detail = FALSE, .stat = character(0))
}
\arguments{
\item{.data}{A data frame}
}
\value{
A tibble
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Enhanced \code{summarise()}.
}
